pyproject.toml
[tool.poetry]
name = "my-site"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "my_site"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
requests = "^2.28.1"
flake8 = "^5.0.4"
flake8-print = "^5.0.0"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
mypy = "^0.981"
pylint = "^2.15.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
